requirements.txt
discord.py
fastapi
motor
uvicorn
pydantic